HHH-9861 Fix wrong mapping of ColumnType FLOAT for H2 database

+++ hibernate-core/src/main/java/org/hibernate/dialect/H2Dialect.java
@@ -118,7 +118,8 @@ public class H2Dialect extends Dialect {
 		registerColumnType( Types.DECIMAL, "decimal($p,$s)" );
 		registerColumnType( Types.NUMERIC, "decimal($p,$s)" );
 		registerColumnType( Types.DOUBLE, "double" );
-		registerColumnType( Types.FLOAT, "float" );
+		// H2 does not have a float type. Instead, it's mapping floats as alias to double
+		registerColumnType( Types.FLOAT, "double" );
 		registerColumnType( Types.INTEGER, "integer" );
 		registerColumnType( Types.LONGVARBINARY, "longvarbinary" );
 		// H2 does define "longvarchar", but it is a simple alias to "varchar"
